add WIP prompts
I was looking at some of anthropics prompts in
https://github.com/anthropics/anthropic-cookbook/tree/main/patterns/agents
which inspired me to try to write longer markdown prompts for our stuff.
I did not get very far before realizing that the specific step I was
trying to write a prompt for should not be done by AI at all,
sidetracking me.

I don't remember how much of the content of those files is my own work
VS from that linked folder. If you build on my work here please take
some time to assess that and make that judgement.
Feel free to consult me about it, I can take a closer look as well.
I'm just to tired to do it today.